requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart
pyyaml
openai>=1.3.0
redis>=4.5.4
qdrant-client>=1.1.1
psycopg2>=2.9.6
markitdown[all]>=0.1.0
httpx>=0.24.0
python-jose[cryptography]
passlib[bcrypt]
python-dotenv